moves code into helper file, makes textarea selection more accurate
